Split the Makefile.common into header and footer components.
For the first time the common Makefile provides something that would
be used on the first pass through the Makefile (PY_EXT_SUFFIX). This
thus needs to be included before the user Makefile, not after. So this
patch splits the single Makefile.common into .header and .footer
versions. There's a compatibility symlink with the old name, so unless
your project is building python extension modules, this isn't a
breaking change