I don't mark all files as secondary
See

  https://www.gnu.org/software/make/manual/make.html#Chained-Rules

This was done to prevent intermediate from being automatically deleted, but I
don't see this actually happening in practice even after this patch. However
marking eerything as secondary had a detrimental effect. If I have a
mrbuild-using Makefile with

  LIB_SOURCES := a.c b.c
  BIN_SOURCES := exe.c

And I

  make && rm *.so* && make

Then exe won't be rebuilt. I could run 'make' again, and then it WOULD be
rebuilt. This is because non-existing secondary files are treated specially.

I simpler demo, is this mrbuild-less Makefile:

  a: b
          cp $< $@
  b:
          seq 5 > $@
  .SECONDARY:

If we "rm a b; touch a; make" then a isn't rebuilt even though b doesn't exist
and we know how to build it. Removing .SECONDARY: fixes that